Remove `AsyncBlockStoreWrite::delete_block`
Why: This function is a footgun that will likely be misused.
Deleting a block is not typically safe because it could have multiple incoming references (and thus deletion converts those into dangling references).
The only way to safely delete a block would be to wholistically enumerate an archive file and collect a list of all referenced blocks.